make Blurbs a container
